Split containerUtils
